в13/src/web/auth.py
"""
Модуль аутентификации для веб-интерфейса
"""
from datetime import datetime, timedelta
from typing import Optional
from jose import JWTError, jwt
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from sqlalchemy.orm import Session

from ..core.config import config
from ..core.database import get_db
from ..core.models import User

# Настройки безопасности
SECRET_KEY = config.SECRET_KEY
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# Контекст для хеширования паролей
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# OAuth2 схема
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Проверка пароля"""
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    """Хеширование пароля"""
    return pwd_context.hash(password)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """Создание JWT токена"""
    to_encode = data.copy()
    
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    
    return encoded_jwt

async def get_current_user(
    token: str = Depends(oauth2_scheme),
    db: Session = Depends(get_db)
) -> User:
    """Получение текущего пользователя из токена"""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    
    # Для простоты возвращаем фиктивного пользователя
    # В реальном приложении нужно искать в БД
    user = User(
        id=1,
        username=username,
        email=f"{username}@example.com",
        is_active=True,
        is_admin=True
    )
    
    return user

class AuthService:
    """Сервис аутентификации"""
    
    def authenticate_user(self, db: Session, username: str, password: str) -> Optional[User]:
        """Аутентификация пользователя"""
        # Для простоты используем хардкод
        # В реальном приложении нужно проверять в БД
        if username == "admin" and password == "admin":
            return User(
                id=1,
                username=username,
                email=f"{username}@example.com",
                is_active=True,
                is_admin=True
            )
        return None
    
    def create_user(self, db: Session, username: str, password: str, email: str) -> User:
        """Создание нового пользователя"""
        hashed_password = get_password_hash(password)
        
        user = User(
            username=username,
            email=email,
            hashed_password=hashed_password,
            is_active=True,
            is_admin=False
        )
        
        db.add(user)
        db.commit()
        db.refresh(user)
        
        return user

# Глобальный экземпляр сервиса
auth_service = AuthService()