в13/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.10.11
websockets==12.0

# Trading
ccxt==4.4.89
pandas==2.1.3
numpy==1.26.2
ta==0.10.2

# Database
PyMySQL==1.1.0
sqlalchemy==2.0.23
redis==5.0.1

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Notifications
python-telegram-bot==20.7

# Utils
python-dotenv==1.0.0
psutil==5.9.6
structlog==23.2.0
APScheduler==3.10.4

# ML (optional)
scikit-learn==1.3.2
